# docs/api-spec.yml
openapi: 3.0.0
info:
  title: Facturas API
  description: |
    REST API for the Facturas multi-tenant invoice management system. Built with Next.js App Router, SQLite (better-sqlite3), and multi-tenant architecture.

    ## Features
    - Multi-tenant invoice management with isolated databases per business
    - Automatic invoice extraction from Gmail attachments (XML/PDF)
    - Confidence scoring and duplicate detection for extracted invoices
    - Full-text search with SQLite FTS5
    - CSV and XLSX export
    - Tags/labels system for invoice categorization
    - API key authentication for public API access
    - Role-based access control (admin/negocio)

    ## Authentication
    - Session-based auth via httpOnly cookies (`session_id`)
    - API key auth for `/api/v1/*` endpoints via `Authorization: Bearer <api_key>`
    - Multi-tenant isolation via `negocio_slug` cookie

  version: 0.1.0
  contact:
    name: Facturas Development Team
  license:
    name: ISC
servers:
  - url: http://localhost:3000
    description: Development server
paths:
  /api/auth/login:
    post:
      summary: Login
      description: Authenticate user with email and password, create session cookie
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
              required: [email, password]
      responses:
        '200':
          description: Login successful, session cookie set
        '401':
          description: Invalid credentials
  /api/auth/register:
    post:
      summary: Register new user
      description: Create a new user account (admin only)
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                nombre:
                  type: string
                role:
                  type: string
                  enum: [admin, negocio]
                negocio_id:
                  type: integer
              required: [email, password, nombre]
      responses:
        '201':
          description: User created
        '400':
          description: Validation error
        '403':
          description: Not authorized
...
  /api/auth/callback:
    get:
      summary: Google OAuth callback
      description: Handle Google OAuth2 callback for Gmail integration
      tags: [Auth]
      parameters:
        - in: query
          name: code
          schema:
            type: string
          description: OAuth2 authorization code
      responses:
        '200':
          description: OAuth tokens stored in cookie
        '400':
          description: Invalid callback
...
  /api/facturas/{id}:
    get:
      summary: Get factura detail
      description: Get a single factura with its line items and attachments
      tags: [Facturas]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Factura with lineas and adjuntos
          content:
            application/json:
              schema:
                type: object
                properties:
                  factura:
                    $ref: '#/components/schemas/Factura'
                  lineas:
                    type: array
                    items:
                      $ref: '#/components/schemas/LineaFactura'
                  adjuntos:
                    type: array
                    items:
                      $ref: '#/components/schemas/Adjunto'
        '404':
          description: Factura not found
    patch:
      summary: Update factura
      description: Update factura fields (currently supports estado changes)
      tags: [Facturas]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                estado:
                  type: string
                  enum: [pendiente, pagada, cancelada]
      responses:
        '200':
          description: Factura updated
        '400':
          description: Invalid estado or no fields
        '404':
          description: Factura not found
...
  /api/extract:
    post:
      summary: Extract invoices from Gmail
      description: |
        Trigger extraction of invoices from Gmail attachments (PDF/XML).
        Downloads attachments via Gmail API, parses them, calculates confidence scores,
        detects duplicates, and stores in the tenant database.
        Sends email notification to admin if errors occur.
      tags: [Extraction]
      responses:
        '200':
          description: Extraction completed with processed/error counts
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  processed:
                    type: integer
                  errors:
                    type: integer
                  details:
                    type: object
                    properties:
                      processed:
                        type: array
                        items:
                          type: object
                          properties:
                            emailId:
                              type: string
                            filename:
                              type: string
                            facturaId:
                              type: integer
                      errors:
                        type: array
                        items:
                          type: object
                          properties:
                            emailId:
                              type: string
                            filename:
                              type: string
                            error:
                              type: string
        '401':
          description: Gmail tokens expired or not authenticated
...
  /api/v1/facturas/{id}:
    get:
      summary: Get factura detail (public API)
      tags: [Public API]
      security:
        - ApiKeyAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Factura detail
        '404':
          description: Not found
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: "Bearer <api_key>"
  schemas:
    Factura:
      type: object
      properties:
        id:
          type: integer
        emisor_nombre:
          type: string
        emisor_nif:
          type: string
          nullable: true
        emisor_direccion:
          type: string
          nullable: true
        emisor_poblacion:
          type: string
          nullable: true
        emisor_provincia:
          type: string
          nullable: true
        emisor_cp:
          type: string
          nullable: true
        emisor_pais:
          type: string
        emisor_email:
          type: string
          nullable: true
        emisor_telefono:
          type: string
          nullable: true
        receptor_nombre:
          type: string
          nullable: true
        receptor_nif:
          type: string
          nullable: true
        receptor_direccion:
          type: string
          nullable: true
        numero_factura:
          type: string
        fecha_emision:
          type: string
          format: date
        fecha_vencimiento:
          type: string
          format: date
          nullable: true
        tipo_documento:
          type: string
          default: factura
        moneda:
          type: string
          default: MXN
        base_imponible:
          type: number
        tipo_iva:
          type: number
          default: 21.0
        cuota_iva:
          type: number
        total:
          type: number
        descuento:
          type: number
          default: 0
        retencion:
          type: number
          default: 0
        neto:
          type: number
          nullable: true
        metodo_pago:
          type: string
          nullable: true
        estado:
          type: string
          enum: [pendiente, pagada, cancelada]
          default: pendiente
        confianza_score:
          type: number
          description: Confidence score 0.0 to 1.0
        confianza_nivel:
          type: string
          enum: [alta, media, baja, error]
          default: alta
        requiere_revision:
          type: integer
          enum: [0, 1]
          default: 0
        revision_notas:
          type: string
          nullable: true
        total_original:
          type: number
          description: Total in original currency (when currency conversion applied)
        moneda_original:
          type: string
          description: Original currency code
        total_convertido:
          type: number
          description: Total converted to negocio default currency
      required: [emisor_nombre, numero_factura, fecha_emision, base_imponible, cuota_iva, total]
    LineaFactura:
      type: object
      properties:
        id:
          type: integer
        factura_id:
          type: integer
        numero_linea:
          type: integer
        descripcion:
          type: string
        cantidad:
          type: number
          default: 1
        precio_unitario:
          type: number
        descuento:
          type: number
          default: 0
        tipo_iva:
          type: number
          default: 21.0
        subtotal:
          type: number
        total:
          type: number
    Adjunto:
      type: object
      properties:
        id:
          type: integer
        factura_id:
          type: integer
        filename:
          type: string
        mime_type:
          type: string
        size:
          type: integer
        attachment_id:
          type: string
        content_hash:
          type: string
...